Fixed a bug in textmate-goto-file
In environments where textmate-cached-project-files would return a
list of files with each containing the full path, the completing read
would display each entry with the full path for each file, and append
the project root to the displayed path, resulting in Emacs opening a
non-existing file in a non-existing folder. The fix should be safe for
platforms where textmate-cached-project-files already returns relative
paths.