Changes find command to output paths relative to project root
Why? If you're in a project-folder that would be otherwise excluded by
*textmate-gf-exclude*, then textmate-goto-file is rendered
non-functional because EVERY file will ignored.  By making find output
relative paths, the ignore regular expression will only operate on files
and directories within the project.

Additionally, `format' will escape double-quotes for us if we use it and
use %S as the placeholder.  This eliminates an unnecessary dependency on
textmate-string-replace.